Separate RawQueries to avoid circular depdendency
