+++ pkg/admin/wire_gen.go
@@ -159,6 +159,9 @@ func newGraphQLHandler(p *deps.RequestProvider) http.Handler {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: store,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := configConfig.FeatureConfig
@@ -462,7 +465,7 @@ func newGraphQLHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service4,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -476,6 +479,7 @@ func newGraphQLHandler(p *deps.RequestProvider) http.Handler {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          store,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -506,7 +510,7 @@ func newGraphQLHandler(p *deps.RequestProvider) http.Handler {
 		TaskQueue: queue,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
