+++ pkg/auth/wire_gen.go
@@ -450,7 +450,7 @@ func newOAuthAuthorizeHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -541,11 +541,15 @@ func newOAuthAuthorizeHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -1038,7 +1042,7 @@ func newOAuthFromWebAppHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -1129,11 +1133,15 @@ func newOAuthFromWebAppHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -1583,7 +1591,7 @@ func newOAuthTokenHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -1681,11 +1689,15 @@ func newOAuthTokenHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -1815,6 +1827,9 @@ func newOAuthRevokeHandler(p *deps.RequestProvider) http.Handler {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: store,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := config.FeatureConfig
@@ -2118,7 +2133,7 @@ func newOAuthRevokeHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -2132,6 +2147,7 @@ func newOAuthRevokeHandler(p *deps.RequestProvider) http.Handler {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          store,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -2211,6 +2227,9 @@ func newOAuthJWKSHandler(p *deps.RequestProvider) http.Handler {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: store,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := config.FeatureConfig
@@ -2511,7 +2530,7 @@ func newOAuthJWKSHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -2525,6 +2544,7 @@ func newOAuthJWKSHandler(p *deps.RequestProvider) http.Handler {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          store,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -2572,6 +2592,9 @@ func newOAuthUserInfoHandler(p *deps.RequestProvider) http.Handler {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: store,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := config.FeatureConfig
@@ -2872,7 +2895,7 @@ func newOAuthUserInfoHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -2886,6 +2909,7 @@ func newOAuthUserInfoHandler(p *deps.RequestProvider) http.Handler {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          store,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -2937,6 +2961,9 @@ func newOAuthEndSessionHandler(p *deps.RequestProvider) http.Handler {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: store,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := config.FeatureConfig
@@ -3236,7 +3263,7 @@ func newOAuthEndSessionHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -3250,6 +3277,7 @@ func newOAuthEndSessionHandler(p *deps.RequestProvider) http.Handler {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          store,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -3636,7 +3664,7 @@ func newOAuthAppSessionTokenHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -3734,11 +3762,15 @@ func newOAuthAppSessionTokenHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -4209,7 +4241,7 @@ func newWebAppLoginHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -4307,11 +4339,15 @@ func newWebAppLoginHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -4801,7 +4837,7 @@ func newWebAppSignupHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -4899,11 +4935,15 @@ func newWebAppSignupHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -5393,7 +5433,7 @@ func newWebAppPromoteHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -5491,11 +5531,15 @@ func newWebAppPromoteHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -5972,7 +6016,7 @@ func newWebAppSelectAccountHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -6070,11 +6114,15 @@ func newWebAppSelectAccountHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -6552,7 +6600,7 @@ func newWebAppSSOCallbackHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -6650,11 +6698,15 @@ func newWebAppSSOCallbackHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -7124,7 +7176,7 @@ func newWechatAuthHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -7222,11 +7274,15 @@ func newWechatAuthHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -7699,7 +7755,7 @@ func newWechatCallbackHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -7797,11 +7853,15 @@ func newWechatCallbackHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -8277,7 +8337,7 @@ func newWebAppEnterLoginIDHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -8375,11 +8435,15 @@ func newWebAppEnterLoginIDHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -8852,7 +8916,7 @@ func newWebAppEnterPasswordHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -8950,11 +9014,15 @@ func newWebAppEnterPasswordHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -9426,7 +9494,7 @@ func newWebAppCreatePasswordHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -9524,11 +9592,15 @@ func newWebAppCreatePasswordHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -10001,7 +10073,7 @@ func newWebAppSetupTOTPHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -10099,11 +10171,15 @@ func newWebAppSetupTOTPHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -10577,7 +10653,7 @@ func newWebAppEnterTOTPHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -10675,11 +10751,15 @@ func newWebAppEnterTOTPHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -11151,7 +11231,7 @@ func newWebAppSetupOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -11249,11 +11329,15 @@ func newWebAppSetupOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -11725,7 +11809,7 @@ func newWebAppEnterOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -11823,11 +11907,15 @@ func newWebAppEnterOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -12301,7 +12389,7 @@ func newWebAppEnterRecoveryCodeHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -12399,11 +12487,15 @@ func newWebAppEnterRecoveryCodeHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -12875,7 +12967,7 @@ func newWebAppSetupRecoveryCodeHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -12973,11 +13065,15 @@ func newWebAppSetupRecoveryCodeHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -13449,7 +13545,7 @@ func newWebAppVerifyIdentityHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -13547,11 +13643,15 @@ func newWebAppVerifyIdentityHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -14026,7 +14126,7 @@ func newWebAppVerifyIdentitySuccessHandler(p *deps.RequestProvider) http.Handler
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -14124,11 +14224,15 @@ func newWebAppVerifyIdentitySuccessHandler(p *deps.RequestProvider) http.Handler
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -14600,7 +14704,7 @@ func newWebAppForgotPasswordHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -14698,11 +14802,15 @@ func newWebAppForgotPasswordHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -15179,7 +15287,7 @@ func newWebAppForgotPasswordSuccessHandler(p *deps.RequestProvider) http.Handler
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -15277,11 +15385,15 @@ func newWebAppForgotPasswordSuccessHandler(p *deps.RequestProvider) http.Handler
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -15753,7 +15865,7 @@ func newWebAppResetPasswordHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -15851,11 +15963,15 @@ func newWebAppResetPasswordHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -16328,7 +16444,7 @@ func newWebAppResetPasswordSuccessHandler(p *deps.RequestProvider) http.Handler 
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -16426,11 +16542,15 @@ func newWebAppResetPasswordSuccessHandler(p *deps.RequestProvider) http.Handler 
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -16902,7 +17022,7 @@ func newWebAppSettingsHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -17000,11 +17120,15 @@ func newWebAppSettingsHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -17495,7 +17619,7 @@ func newWebAppSettingsIdentityHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -17593,11 +17717,15 @@ func newWebAppSettingsIdentityHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -18071,7 +18199,7 @@ func newWebAppSettingsBiometricHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -18169,11 +18297,15 @@ func newWebAppSettingsBiometricHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -18646,7 +18778,7 @@ func newWebAppSettingsMFAHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -18744,11 +18876,15 @@ func newWebAppSettingsMFAHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -19230,7 +19366,7 @@ func newWebAppSettingsTOTPHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -19328,11 +19464,15 @@ func newWebAppSettingsTOTPHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -19805,7 +19945,7 @@ func newWebAppSettingsOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -19903,11 +20043,15 @@ func newWebAppSettingsOOBOTPHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -20380,7 +20524,7 @@ func newWebAppSettingsRecoveryCodeHandler(p *deps.RequestProvider) http.Handler 
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -20478,11 +20622,15 @@ func newWebAppSettingsRecoveryCodeHandler(p *deps.RequestProvider) http.Handler 
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -20956,7 +21104,7 @@ func newWebAppSettingsSessionsHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -21054,11 +21202,15 @@ func newWebAppSettingsSessionsHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -21537,7 +21689,7 @@ func newWebAppForceChangePasswordHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -21635,11 +21787,15 @@ func newWebAppForceChangePasswordHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -22112,7 +22268,7 @@ func newWebAppSettingsChangePasswordHandler(p *deps.RequestProvider) http.Handle
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -22210,11 +22366,15 @@ func newWebAppSettingsChangePasswordHandler(p *deps.RequestProvider) http.Handle
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -22687,7 +22847,7 @@ func newWebAppForceChangeSecondaryPasswordHandler(p *deps.RequestProvider) http.
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -22785,11 +22945,15 @@ func newWebAppForceChangeSecondaryPasswordHandler(p *deps.RequestProvider) http.
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -23262,7 +23426,7 @@ func newWebAppSettingsChangeSecondaryPasswordHandler(p *deps.RequestProvider) ht
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -23360,11 +23524,15 @@ func newWebAppSettingsChangeSecondaryPasswordHandler(p *deps.RequestProvider) ht
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -23837,7 +24005,7 @@ func newWebAppUserDisabledHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -23935,11 +24103,15 @@ func newWebAppUserDisabledHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -24411,7 +24583,7 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -24509,11 +24681,15 @@ func newWebAppLogoutHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -25005,7 +25181,7 @@ func newWebAppReturnHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -25103,11 +25279,15 @@ func newWebAppReturnHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -25579,7 +25759,7 @@ func newWebAppErrorHandler(p *deps.RequestProvider) http.Handler {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -25677,11 +25857,15 @@ func newWebAppErrorHandler(p *deps.RequestProvider) http.Handler {
 		Clock: clockClock,
 	}
 	commands := &user.Commands{
-		Raw:          rawCommands,
+		RawCommands:  rawCommands,
 		Events:       eventService,
 		Verification: verificationService,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
@@ -26112,6 +26296,9 @@ func newSessionMiddleware(p *deps.RequestProvider) httproute.Middleware {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := config.FeatureConfig
@@ -26384,7 +26571,7 @@ func newSessionMiddleware(p *deps.RequestProvider) httproute.Middleware {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -26398,6 +26585,7 @@ func newSessionMiddleware(p *deps.RequestProvider) httproute.Middleware {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
