+++ pkg/lib/authn/user/queries.go
@@ -4,8 +4,6 @@ import (
 	"github.com/authgear/authgear-server/pkg/api/model"
 	"github.com/authgear/authgear-server/pkg/lib/authn/authenticator"
 	"github.com/authgear/authgear-server/pkg/lib/authn/identity"
-	"github.com/authgear/authgear-server/pkg/lib/infra/db"
-	"github.com/authgear/authgear-server/pkg/util/graphqlutil"
 )
 
 type IdentityService interface {
@@ -21,6 +19,7 @@ type VerificationService interface {
 }
 
 type Queries struct {
+	*RawQueries
 	Store          store
 	Identities     IdentityService
 	Authenticators AuthenticatorService
@@ -49,35 +48,4 @@ func (p *Queries) Get(id string) (*model.User, error) {
 	}
 
 	return newUserModel(user, identities, authenticators, isVerified), nil
-}
-
-func (p *Queries) GetRaw(id string) (*User, error) {
-	return p.Store.Get(id)
-}
-
-func (p *Queries) GetManyRaw(ids []string) ([]*User, error) {
-	return p.Store.GetByIDs(ids)
-}
-
-func (p *Queries) Count() (uint64, error) {
-	return p.Store.Count()
-}
-
-func (p *Queries) QueryPage(sortOption SortOption, pageArgs graphqlutil.PageArgs) ([]model.PageItemRef, error) {
-	users, offset, err := p.Store.QueryPage(sortOption, pageArgs)
-	if err != nil {
-		return nil, err
-	}
-
-	var models = make([]model.PageItemRef, len(users))
-	for i, u := range users {
-		pageKey := db.PageKey{Offset: offset + uint64(i)}
-		cursor, err := pageKey.ToPageCursor()
-		if err != nil {
-			return nil, err
-		}
-
-		models[i] = model.PageItemRef{ID: u.ID, Cursor: cursor}
-	}
-	return models, nil
 }
