+++ pkg/resolver/wire_gen.go
@@ -183,6 +183,9 @@ func newSessionMiddleware(p *deps.RequestProvider) httproute.Middleware {
 		SQLBuilder:  sqlBuilderApp,
 		SQLExecutor: sqlExecutor,
 	}
+	rawQueries := &user.RawQueries{
+		Store: userStore,
+	}
 	authenticationConfig := appConfig.Authentication
 	identityConfig := appConfig.Identity
 	featureConfig := config.FeatureConfig
@@ -455,7 +458,7 @@ func newSessionMiddleware(p *deps.RequestProvider) httproute.Middleware {
 		Authenticators:  service3,
 		Verification:    verificationService,
 		MFA:             mfaService,
-		Users:           rawCommands,
+		UserCommands:    rawCommands,
 		PasswordHistory: historyStore,
 		OAuth:           authorizationStore,
 		IDPSessions:     idpsessionManager,
@@ -469,6 +472,7 @@ func newSessionMiddleware(p *deps.RequestProvider) httproute.Middleware {
 		Coordinator: coordinator,
 	}
 	queries := &user.Queries{
+		RawQueries:     rawQueries,
 		Store:          userStore,
 		Identities:     identityFacade,
 		Authenticators: authenticatorFacade,
